[YS-479] fix: popover 내부에서 저장되지 않은 상태는 반영되지 않도록 처리
