+++ src/app/home/components/ExperimentPostContainer/FilterContainer/AreaFilter/components/AreaFilterBottomSheet/AreaFilterBottomSheet.tsx
@@ -30,7 +30,6 @@ const hasSelectedAreaAll = (selectedAreas: Record<string, boolean>) => {
 
 interface AreaFilterBottomSheetProps {
   onChange: (filters: Record<string, string | string[] | number | null>) => void;
-  onReset: () => void;
   onClose: () => void;
   initialRegion?: RegionType;
   initialAreas?: AreaType[];
@@ -38,7 +37,6 @@ interface AreaFilterBottomSheetProps {
 
 const AreaFilterBottomSheet = ({
   onChange,
-  onReset,
   onClose,
   initialRegion,
   initialAreas,
@@ -79,7 +77,6 @@ const AreaFilterBottomSheet = ({
   const handleReset = () => {
     setSelectedRegion(null);
     setSelectedAreas({});
-    onReset();
   };
 
   const handleConfirm = () => {
