+++ src/app/home/components/ExperimentPostContainer/FilterContainer/AreaFilter/components/FooterButtonContainer/FooterButtonContainer.tsx
@@ -2,7 +2,6 @@ import {
   buttonContainer,
   footerButton,
   footerButtonContainer,
-  footerContainer,
   infoText,
   infoTextContainer,
 } from './FooterButtonContainer.css';
@@ -22,26 +21,24 @@ const FooterButtonContainer = ({
   isValidSaveButton,
 }: FooterButtonContainerProps) => {
   return (
-    <div className={footerContainer}>
-      <div className={footerButtonContainer}>
-        <div className={infoTextContainer}>
-          <Icon icon="Information" width={16} height={16} />
-          <span className={infoText}>최대 5개까지 선택할 수 있어요</span>
-        </div>
-        <div className={buttonContainer}>
-          <Button variant="secondary" size="small" className={footerButton} onClick={handleReset}>
-            초기화
-          </Button>
-          <Button
-            variant="primary"
-            size="small"
-            onClick={handleSave}
-            className={footerButton}
-            disabled={!isValidSaveButton}
-          >
-            저장
-          </Button>
-        </div>
+    <div className={footerButtonContainer}>
+      <div className={infoTextContainer}>
+        <Icon icon="Information" width={16} height={16} />
+        <span className={infoText}>최대 5개까지 선택할 수 있어요</span>
+      </div>
+      <div className={buttonContainer}>
+        <Button variant="secondary" size="small" className={footerButton} onClick={handleReset}>
+          초기화
+        </Button>
+        <Button
+          variant="primary"
+          size="small"
+          onClick={handleSave}
+          className={footerButton}
+          disabled={!isValidSaveButton}
+        >
+          저장
+        </Button>
       </div>
     </div>
   );
